refactored users.js
